[Fix]: #37 Coderabbit AI 관련 PR 수정
