+++ src/main/java/ssurent/ssurentbe/domain/users/service/UserPenaltyCommandService.java
@@ -45,36 +45,4 @@ public class UserPenaltyCommandService {
                 .orElseThrow(() -> new GeneralException(ErrorStatus.PENALTY_NOT_FOUND));
         userPenaltyLogRepository.delete(userPenaltyLog);
     }
-
-    @Transactional
-    public void bulkUpsertUsers(List<AdminBulkUserUpdateRequest> requests) {
-
-        try{
-            for (AdminBulkUserUpdateRequest req : requests) {
-                Optional<Users> existing = userRepository.findByStudentNum(req.studentNum());
-
-                if (existing.isPresent()) {
-                    Users user = existing.get();
-                    user.updateInfo(req.name(), req.phoneNum());
-                } else {
-                    Users user = Users.builder()
-                            .studentNum(req.studentNum())
-                            .name(req.name())
-                            .phoneNum(req.phoneNum())
-                            .password(
-                                    req.phoneNum().isBlank()
-                                            ? req.studentNum()
-                                            : req.phoneNum()
-                            )
-                            .role(Role.NORMAL)
-                            .deleted(false)
-                            .build();
-                    userRepository.save(user);
-                }
-            }
-        }
-        catch (Exception e){
-            throw new GeneralException(ErrorStatus.USER_UPSERT_FAILED);
-        }
-    }
 }
